feat: Enhance Forward+ rendering with improved light culling and shader adjustments

+++ OloEditor/assets/shaders/PBR_MultiLight_Skinned.glsl
@@ -10,7 +10,7 @@
 layout(location = 0) in vec3 a_Position;
 layout(location = 1) in vec3 a_Normal;
 layout(location = 2) in vec2 a_TexCoord;
-layout(location = 3) in vec4 a_BoneIDs;
+layout(location = 3) in ivec4 a_BoneIDs;
 layout(location = 4) in vec4 a_BoneWeights;
 
 // Camera UBO (binding 0)
@@ -48,7 +48,7 @@ void main()
     mat4 boneTransform = mat4(0.0);
     for (int i = 0; i < 4; ++i)
     {
-        int boneID = int(a_BoneIDs[i]);
+        int boneID = a_BoneIDs[i];
         if (boneID >= 0 && boneID < 100)
         {
             boneTransform += u_BoneTransforms[boneID] * a_BoneWeights[i];
